# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Prototype Hono Drizzle API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /messages/{id}:
    delete:
      summary: Delete a message
      description: Soft-delete a message. Only the sender or conversation admin can delete.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Message successfully deleted
        '403':
          description: Forbidden - user is not authorized to delete this message
        '404':
          description: Message not found
  /messages/{id}/reactions:
    get:
      summary: Get reactions for a message
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of reactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reaction'
        '404':
          description: Message not found
    post:
      summary: Add reaction to message
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionRequest'
      responses:
        '201':
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'
  /messages/{id}/reactions/{emoji}:
    delete:
      summary: Remove reaction from message
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: emoji
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reaction removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'
...
  /messages/{id}/bookmarks:
    post:
      summary: Bookmark a message
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkRequest'
      responses:
        '201':
          description: Message bookmarked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkResponse'
    delete:
      summary: Remove bookmark from a message
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bookmark removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnbookmarkResponse'
  /users:
    get:
      summary: List all users (development only)
      description: Get a list of all users. Only available in development mode.
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - only available in development mode
    post:
      summary: Create user (development only)
      description: Creates a new user. Only available in development mode.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - only available in development mode
  /users/{userId}:
    get:
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /users/{userId}/bookmarks:
    get:
      summary: List bookmarks for a user
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookmarkListItem'
  /users/login:
    post:
      summary: Login user by ID alias
      description: Authenticate and retrieve user information by their unique ID alias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - invalid idAlias format
        '404':
          description: User not found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  schemas:
...
    CreateUserRequest:
      type: object
      properties:
        idAlias:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]*[a-z0-9]$'
          minLength: 3
          maxLength: 30
          description: 'Unique human-readable identifier. Only lowercase letters, numbers, dots, underscores, and hyphens allowed. No spaces.'
        name:
          type: string
          minLength: 1
        avatarUrl:
          type: string
          nullable: true
      required:
        - idAlias
        - name
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        idAlias:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]*[a-z0-9]$'
          minLength: 3
          maxLength: 30
          description: 'Unique human-readable identifier for login and display. Must start and end with lowercase letter or number. Only lowercase letters, numbers, dots, underscores, and hyphens allowed. No spaces.'
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - idAlias
        - name
        - createdAt
...
    Reaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        emoji:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - messageId
        - userId
        - emoji
        - createdAt
...
    ReactionRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        emoji:
          type: string
      required:
        - userId
        - emoji
...
    Bookmark:
      type: object
      properties:
        id:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - messageId
        - userId
        - createdAt
    BookmarkRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
      required:
        - userId
    BookmarkResponse:
      type: object
      properties:
        status:
          type: string
          enum: [bookmarked]
        bookmark:
          $ref: '#/components/schemas/Bookmark'
      required:
        - status
        - bookmark
    UnbookmarkResponse:
      type: object
      properties:
        status:
          type: string
          enum: [unbookmarked]
      required:
        - status
    BookmarkListItem:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        text:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        messageCreatedAt:
          type: string
          format: date-time
      required:
        - messageId
        - conversationId
        - createdAt
        - messageCreatedAt
    LoginRequest:
      type: object
      properties:
        idAlias:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]*[a-z0-9]$'
          minLength: 3
          maxLength: 30
      required:
        - idAlias
...